requirements.txt
streamlit
pillow
psycopg2-binary
python-dotenv
streamlit-calendar
